molecule/driver/docker.py
#  Copyright (c) 2015-2016 Cisco Systems, Inc.
#
#  Permission is hereby granted, free of charge, to any person obtaining a copy
#  of this software and associated documentation files (the "Software"), to
#  deal in the Software without restriction, including without limitation the
#  rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
#  sell copies of the Software, and to permit persons to whom the Software is
#  furnished to do so, subject to the following conditions:
#
#  The above copyright notice and this permission notice shall be included in
#  all copies or substantial portions of the Software.
#
#  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
#  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
#  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
#  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
#  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
#  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
#  DEALINGS IN THE SOFTWARE.

import collections

from molecule.driver import base


class Docker(base.Base):
    def __init__(self, config):
        super(Docker, self).__init__(config)

    @property
    def testinfra_options(self):
        return {'connection': 'docker'}

    @property
    def inventory(self):
        # instance-1-default ansible_connection=docker
        # instance-2-default ansible_connection=docker
        host_options = 'ansible_connection=docker'
        dd = collections.defaultdict(list)
        for d in self._config.platforms:
            name = '{}-{}'.format(d['name'], self._config.scenario.name)
            dd[name].append(host_options)

        return dict(dd)
